Added homepage and corrected login functionality
