+++ client/package.json
@@ -7,13 +7,13 @@
     "@devexpress/dx-react-chart": "^2.7.6",
     "@devexpress/dx-react-chart-material-ui": "^2.7.6",
     "@devexpress/dx-react-core": "^2.7.6",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.58",
     "@material-ui/pickers": "^3.3.10",
-    "@fortawesome/fontawesome-svg-core": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
     "@react-oauth/google": "^0.9.0",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
@@ -34,10 +34,11 @@
     "react-nice-dates": "^3.1.0",
     "react-progress-button": "^5.1.0",
     "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "4.0.3",
     "react-simple-snackbar": "^1.1.11",
     "react-tiny-fab": "^4.0.3",
+    "react-toastify": "^9.0.0",
     "recharts": "^2.0.9",
     "redux": "^4.1.0",
     "redux-thunk": "^2.3.0",
