+++ client/src/components/Footer/Footer.css
@@ -143,6 +143,26 @@
     fill: #1A8EFD;
     border: 2px solid #1A8EFD;
 }
+
+.text-appointment-btn {
+    padding: 14px 20px;
+    color: white;
+    border: 1px solid transparent;
+    border-radius: 28px;
+    outline: transparent;
+    background-color: #1A8EFD;
+    font-size: 18px;
+    font-family: 'Rubik', sans-serif;
+    letter-spacing: .8px;
+    cursor: pointer;
+    transition: all .4s ease;
+}
+
+.text-appointment-btn:hover {
+    color: #1A8EFD;
+    background-color: transparent;
+    border: 1px solid #1A8EFD;
+}
 /* Responsive */
 @media screen and (max-width: 700px) {
     .footer-container {
@@ -167,5 +187,8 @@
     }
     .ft-social-links {
         margin: 16px 0 0;
+    }
+    .text-appointment-btn {
+        font-size: 16px;
     }
 }
