+++ client/src/components/Home/Home.module.css
@@ -15,6 +15,15 @@
   height: 70px;
 }
 
+.header2 {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 60px;
+  border-bottom: 1px solid rgb(218, 218, 218);
+  padding-left: 20px;
+}
+
 .login {
   border: 1px solid rgb(51, 51, 216);
   padding: 12px 30px;
