+++ client/src/components/Homepage/Homepage.js
@@ -1,0 +1,30 @@
+import { useNavigate } from "react-router-dom";
+import Navbar from "../Navbar/Navbar";
+import Hero from "../Hero/Hero";
+import Info from "../Info/Info";
+import About from "../About/About";
+// import Reviews from "../Reviews/Reviews";
+import Footer from "../Footer/Footer";
+
+const Homepage = () => {
+
+    const navigate = useNavigate();
+    const user = JSON.parse(localStorage.getItem('profile'))
+
+    if(!user) {
+        navigate('/login')
+    }
+
+    return (
+        <div className="homepage">
+            <Navbar />
+            <Hero />
+            <Info />
+            <About />
+            {/* <Reviews /> */}
+            <Footer />
+        </div>
+    )
+}
+
+export default Homepage
