+++ client/src/components/NavBar/NavBar.js
@@ -1,74 +1,138 @@
-import React, { useState, useEffect } from 'react'
-import { useLocation, Link } from 'react-router-dom'
+import React, { useState } from "react";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {
+  faCalendarDays,
+  faBars,
+  faXmark,
+  faSignOut,
+} from "@fortawesome/free-solid-svg-icons";
+import "./Navbar.css";
+import { Link } from "react-router-dom";
+import { toast } from "react-toastify";
+import { useNavigate } from "react-router-dom";
 
-const NavBar = () => {
-
-  const location = useLocation()
+function Navbar() {
+  const [nav, setNav] = useState(false);
+  const [isButtonDisabled, setIsButtonDisabled] = useState(false);
   const [user, setUser] = useState(JSON.parse(localStorage.getItem('profile')))
 
-  useEffect(() => {
-    setUser(JSON.parse(localStorage.getItem('profile')))
-  }, [location])
+  const navigate = useNavigate();
 
-    if(!user) return null
+  const openNav = () => {
+    setNav(!nav);
+  };
 
-    return (
-        <div>
-           <nav className="navbar">
-  <ul className="navbar-nav">
-    <li className="logo">
-      <a href="/dashboard" className="nav-link">
-        <span className="link-text logo-text"><img style={{width: '50px'}} src="https://i.postimg.cc/hGZKzdkS/logo.png" alt="arc-invoice" /></span>
-        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-arrow-right"><line x1="5" y1="12" x2="19" y2="12"></line><polyline points="12 5 19 12 12 19"></polyline></svg>
-      </a>
-    </li>
-    <li className="nav-item">
-      <Link to="/dashboard" className="nav-link">
-      <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-pie-chart"><path d="M21.21 15.89A10 10 0 1 1 8 2.83"></path><path d="M22 12A10 10 0 0 0 12 2v10z"></path></svg>
-        <span className="link-text">Dashboard</span>
-      </Link>
-    </li>
-    
-    <li className="nav-item">
-      <a href="/invoice" className="nav-link">
-      <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-plus"><line x1="12" y1="5" x2="12" y2="19"></line><line x1="5" y1="12" x2="19" y2="12"></line></svg>
-        <span className="link-text">Create</span>
-      </a>
-    </li>
-    <li className="nav-item">
-      <a href="/invoices" className="nav-link">
-      <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-layers"><polygon points="12 2 2 7 12 12 22 7 12 2"></polygon><polyline points="2 17 12 22 22 17"></polyline><polyline points="2 12 12 17 22 12"></polyline></svg>
-        <span className="link-text">Invoices</span>
-      </a>
-    </li>
-    <li className="nav-item">
-      <a href="/customers" className="nav-link">
-      <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-users"><path d="M17 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2"></path><circle cx="9" cy="7" r="4"></circle><path d="M23 21v-2a4 4 0 0 0-3-3.87"></path><path d="M16 3.13a4 4 0 0 1 0 7.75"></path></svg>
-        <span className="link-text">Customers</span>
-      </a>
-    </li>
-    {/* <li className="nav-item">
-      <a href="#" className="nav-link">
-        <svg aria-hidden="true" focusable="false" data-prefix="fad" data-icon="space-shuttle" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 640 512" className="svg-inline--fa fa-space-shuttle fa-w-20 fa-5x">
-          <g className="fa-group">
-            <path fill="currentColor" d="M32 416c0 35.35 21.49 64 48 64h16V352H32zm154.54-232h280.13L376 168C243 140.59 222.45 51.22 128 34.65V160h18.34a45.62 45.62 0 0 1 40.2 24zM32 96v64h64V32H80c-26.51 0-48 28.65-48 64zm114.34 256H128v125.35C222.45 460.78 243 371.41 376 344l90.67-16H186.54a45.62 45.62 0 0 1-40.2 24z" className="fa-secondary" />
-            <path fill="currentColor" d="M592.6 208.24C559.73 192.84 515.78 184 472 184H186.54a45.62 45.62 0 0 0-40.2-24H32c-23.2 0-32 10-32 24v144c0 14 8.82 24 32 24h114.34a45.62 45.62 0 0 0 40.2-24H472c43.78 0 87.73-8.84 120.6-24.24C622.28 289.84 640 272 640 256s-17.72-33.84-47.4-47.76zM488 296a8 8 0 0 1-8-8v-64a8 8 0 0 1 8-8c31.91 0 31.94 80 0 80z" className="fa-primary" />
-          </g>
-        </svg>
-        <span className="link-text">Shuttle</span>
-      </a>
-    </li> */}
-    <li className="nav-item" id="themeButton">
-      <a href="/settings" className="nav-link">
-      <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-settings"><circle cx="12" cy="12" r="3"></circle><path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1 0 2.83 2 2 0 0 1-2.83 0l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-2 2 2 2 0 0 1-2-2v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83 0 2 2 0 0 1 0-2.83l.06-.06a1.65 1.65 0 0 0 .33-1.82 1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1-2-2 2 2 0 0 1 2-2h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 0-2.83 2 2 0 0 1 2.83 0l.06.06a1.65 1.65 0 0 0 1.82.33H9a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 2-2 2 2 0 0 1 2 2v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 0 2 2 0 0 1 0 2.83l-.06.06a1.65 1.65 0 0 0-.33 1.82V9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 2 2 2 2 0 0 1-2 2h-.09a1.65 1.65 0 0 0-1.51 1z"></path></svg>
-        <span className="link-text">Settings</span>
-      </a>
-    </li>
-  </ul>
-</nav>
- 
+  const handleChatBtnClick = () => {
+    if (!isButtonDisabled) {
+      toast.info("Experiencing high traffic, Please wait a moment.", {
+        position: toast.POSITION.TOP_CENTER,
+        onOpen: () => setIsButtonDisabled(true),
+        onClose: () => setIsButtonDisabled(false),
+      });
+    }
+  };
+
+  const handleLogoutBtn = () => {
+      navigate('/')
+      setUser(null)
+      localStorage.clear()
+  }
+
+  return (
+    <div className="navbar-section">
+      <h1 className="navbar-title">
+        <Link to="/homepage">
+          AI <span className="navbar-sign">Interviewer</span>
+        </Link>
+      </h1>
+
+      {/* Desktop */}
+      <ul className="navbar-items">
+        <li>
+          <Link to="/homepage" className="navbar-links">
+            Home
+          </Link>
+        </li>
+        <li>
+          <a href="#services" className="navbar-links">
+            Services
+          </a>
+        </li>
+        <li>
+          <a href="#about" className="navbar-links">
+            About
+          </a>
+        </li>
+        <li>
+          <a href="#reviews"  className="navbar-links">
+            Reviews
+          </a>
+        </li>
+      </ul>
+
+      {/* <button
+        className="navbar-btn"
+        type="button"
+        disabled={isButtonDisabled}
+        onClick={handleChatBtnClick}
+      >
+        <FontAwesomeIcon icon={faCalendarDays} /> Schedule an interview
+      </button> */}
+      <button
+        className="navbar-btn"
+        type="button"
+        disabled={isButtonDisabled}
+        onClick={handleLogoutBtn}
+      >
+        <FontAwesomeIcon icon={faSignOut} /> 
+      </button>
+
+
+      {/* Mobile */}
+      <div className={`mobile-navbar ${nav ? "open-nav" : ""}`}>
+        <div onClick={openNav} className="mobile-navbar-close">
+          <FontAwesomeIcon icon={faXmark} className="hamb-icon" />
         </div>
-    )
+
+        <ul className="mobile-navbar-links">
+          <li>
+            <Link onClick={openNav} to="/">
+              Home
+            </Link>
+          </li>
+          <li>
+            <a onClick={openNav} href="#services">
+              Services
+            </a>
+          </li>
+          <li>
+            <a onClick={openNav} href="#about">
+              About
+            </a>
+          </li>
+          <li>
+            <a onClick={openNav} href="#reviews">
+              Reviews
+            </a>
+          </li>
+          {/* <li>
+            <a onClick={openNav} href="#contact">
+              Contact
+            </a>
+          </li> */}
+        </ul>
+      </div>
+
+      {/* Hamburger Icon */}
+      <div className="mobile-nav">
+        <FontAwesomeIcon
+          icon={faBars}
+          onClick={openNav}
+          className="hamb-icon"
+        />
+      </div>
+    </div>
+  );
 }
 
-export default NavBar
+export default Navbar;
